cast window start and end times as int

+++ visual_behavior/visualization/ophys/summary_figures.py
@@ -175,7 +175,7 @@ def get_xticks_xticklabels(trace, frame_rate, interval_sec=1, window=None):
 
     :return: xticks, xticklabels = xticks in frames corresponding to timepoints in the trace, xticklabels in seconds
     """
-    interval_frames = interval_sec * frame_rate
+    interval_frames = int(interval_sec * frame_rate)
     n_frames = len(trace)
     n_sec = n_frames / frame_rate
     xticks = np.arange(0, n_frames + 1, interval_frames)
